requirements.txt
litellm==1.48.0
openai==1.52.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
httpcore==1.0.5
wikipedia-api==0.6.0
wolframalpha==5.0.0
bitsandbytes==0.43.1
datasets==4.0.0
deepspeed==0.13.2
einops==0.7.0
isort==5.13.2
jsonlines==4.0.0
loralib==0.1.2
optimum==1.18.0
peft==0.11.1
torchmetrics==1.4.0
tqdm==4.66.4
termcolor==2.4.0
wandb==0.17.5
pyyaml==6.0.1
google-search-results==2.4.2
requests-cache==1.2.0
openmeteo-requests==1.7.0
retry-requests==2.0.0
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
dotmap==1.3.30
s3fs==2024.6.1
setuptools==80.9.0
sphere-engine==0.27
simplejson==3.19.2
pandas==2.3.1
pyOpenSSL==24.1.0
pydantic==2.9
standard-imghdr==3.13.0